tilegrid iob: factor out generic code
