+++ fuzzers/005-tilegrid/fuzzaddr/generate.py
@@ -1,9 +1,8 @@
 #!/usr/bin/env python3
 
-import bitsmaker
+from prjxray import bitsmaker
 
-
-def run(bits_fn, design_fn, fnout, verbose=False):
+def run(bits_fn, design_fn, fnout, oneval, verbose=False):
     # Raw: IOB_X0Y101 00020027_003_03
     metastr = "DFRAME:27.DWORD:3.DBIT:3"
 
@@ -12,15 +11,9 @@ def run(bits_fn, design_fn, fnout, verbose=False):
     f.readline()
     for l in f:
         l = l.strip()
-        port, site, tile, pin, val = l.split(',')
-        '''
-        PULLTYPE    28  29  30
-        NONE                X
-        KEEPER      X       X
-        PULLDOWN
-        PULLUP          X   X
-        '''
-        tags["%s.%s" % (tile, metastr)] = val == "KEEPER"
+        # Additional values reserved / for debugging
+        tile, val = l.split(',')[0:2]
+        tags["%s.%s" % (tile, metastr)] = val == oneval
 
     bitsmaker.write(bits_fn, fnout, tags)
 
@@ -35,9 +28,10 @@ def main():
     parser.add_argument("--verbose", action="store_true", help="")
     parser.add_argument("--design", default="design.csv", help="")
     parser.add_argument("--fnout", default="/dev/stdout", help="")
+    parser.add_argument("--oneval", required=True, help="")
     args = parser.parse_args()
 
-    run(args.bits_file, args.design, args.fnout, args.verbose)
+    run(args.bits_file, args.design, args.fnout, oneval=args.oneval, verbose=args.verbose)
 
 
 if __name__ == "__main__":
